# OpenAPI.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Tutorial Management API",
    "description": "This API exposes endpoints to manage tutorials.",
    "termsOfService": "https://www.bezkoder.com/terms",
    "contact": {
      "name": "My Name",
      "url": "https://www.fct.unl.pt",
      "email": "myemail@gmail.com"
    },
    "license": {
      "name": "MIT License",
      "url": "https://choosealicense.com/licenses/mit/"
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080/",
      "description": "Server URL in Development environment"
    },
    {
      "url": "https://api.example.com/",
      "description": "Server URL in Production environment"
    }
  ],
  "paths": {
    "/owners/{ownerId}/bookings/{bookingId}": {
      "put": {
        "tags": [
          "owner-controller"
        ],
        "operationId": "changeBookingStatus",
        "parameters": [
          {
            "name": "ownerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "bookingId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "string"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Status changed successfully."
          },
          "400": {
            "description": "Invalid decision.",
            "content": {}
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Booking not found; Owner not found.",
            "content": {}
          },
          "409": {
            "description": "Booking is not Under Consideration.",
            "content": {}
          }
        }
      }
    },
...
    "/owners/": {
      "post": {
        "tags": [
          "owner-controller"
        ],
        "operationId": "createOwner",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "409": {
            "description": "Conflict"
          }
        }
      }
    },
    "/clients/{username}/{bookingId}/checkedOut": {
      "post": {
        "tags": [
          "client-controller"
        ],
        "operationId": "checkOut",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "bookingId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Guest Checked Out"
          },
          "400": {
            "description": "Client has not checked in yet or booking has been rejected; Check-out day hasn't arrived yet.",
            "content": {}
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Booking not found; Client not found.",
            "content": {}
          }
        }
      }
    },
    "/clients/{username}/{bookingId}/checkedIn": {
      "post": {
        "tags": [
          "client-controller"
        ],
        "operationId": "checkIn",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "bookingId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Guest Checked In"
          },
          "400": {
            "description": "Booking has not been accepted yet or has been rejected; Check-in day hasn't arrived yet.",
            "content": {}
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Booking not found; Client not found",
            "content": {}
          }
        }
      }
    },
    "/clients/{username}/{bookingId}/cancelled": {
      "post": {
        "tags": [
          "client-controller"
        ],
        "operationId": "cancelBooking",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "bookingId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Booking canceled successfully"
          },
          "400": {
            "description": "Booking cant be cancelled",
            "content": {}
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Booking not found; Client not found",
            "content": {}
          }
        }
      }
    },
    "/clients/": {
      "post": {
        "tags": [
          "client-controller"
        ],
        "operationId": "createClient",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClientInfoDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Client was successfully created"
          },
          "400": {
            "description": "Invalid Client data",
            "content": {}
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "409": {
            "description": "Username unavailable"
          }
        }
      }
    },
...
    "/owners/{ownerId}": {
      "get": {
        "tags": [
          "owner-controller"
        ],
        "operationId": "getOwner",
        "parameters": [
          {
            "name": "ownerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Owner."
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Owner not found.",
            "content": {}
          },
          "409": {
            "description": "Conflict"
          }
        }
      }
    },
    "/owners/{ownerId}/apartments": {
      "get": {
        "tags": [
          "owner-controller"
        ],
        "operationId": "getOwnerApartments",
        "parameters": [
          {
            "name": "ownerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Owner Apartments.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ApartmentInfoDTO"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Owner not found.",
            "content": {}
          },
          "409": {
            "description": "Conflict"
          }
        }
      }
    },
    "/owners/{ownerId}/apartments-bookings": {
      "get": {
        "tags": [
          "owner-controller"
        ],
        "operationId": "getOwnerBookings",
        "parameters": [
          {
            "name": "ownerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Bookings for each apartment owned.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BookingInfoDTO"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Owner not found.",
            "content": {}
          },
          "409": {
            "description": "Conflict"
          }
        }
      }
    },
    "/clients/{username}": {
      "get": {
        "tags": [
          "client-controller"
        ],
        "operationId": "getClient",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Client that is in the system",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientInfoDTO"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Client not found.",
            "content": {}
          }
        }
      }
    },
    "/clients/{username}/reviews": {
      "get": {
        "tags": [
          "client-controller"
        ],
        "operationId": "getClientReviews",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of client reviews",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReviewInfoDTO"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Client not found.",
            "content": {}
          }
        }
      }
    },
    "/clients/{username}/bookings": {
      "get": {
        "tags": [
          "client-controller"
        ],
        "operationId": "getClientBookings",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of client bookings",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BookingInfoDTO"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Client not found.",
            "content": {}
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ReviewInfoDTO": {
        "required": [
          "clientName"
        ],
        "type": "object",
        "properties": {
          "clientName": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "rating": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "ClientInfoDTO": {
        "required": [
          "email",
          "name",
          "phoneNumber"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "ApartmentInfoDTO": {
        "required": [
          "amenities",
          "description",
          "id",
          "location",
          "name",
          "pictures",
          "price"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "location": {
            "type": "string"
          },
          "amenities": {
            "type": "string"
          },
          "pictures": {
            "type": "string"
          },
          "price": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "BookingInfoDTO": {
        "required": [
          "apartmentInfo",
          "checkIn",
          "checkOut",
          "clientName",
          "guests",
          "id",
          "state"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "clientName": {
            "type": "string"
          },
          "checkIn": {
            "type": "string",
            "format": "date"
          },
          "checkOut": {
            "type": "string",
            "format": "date"
          },
          "guests": {
            "type": "integer",
            "format": "int64"
          },
          "state": {
            "type": "string"
          },
          "apartmentInfo": {
            "$ref": "#/components/schemas/ApartmentInfoDTO"
          }
        }
      },
...
    }
  }
}